+++ Versions/Temporary/Engine/Sources/SceneB2/WaterController.cpp
@@ -249,7 +249,7 @@ void CWaterController::ProcessWavesDistribution( CArray2D<CPtr<SWaterNode> > *pW
 			float fAmpl;
 			if ( bIsCurWave )
 			{
-				fAmpl = Clamp( fabs( ( fCurWaveStart + fCurWaveLen ) - fCurOffset * 2.0f ) / ( fCurWaveLen - fCurWaveStart ), 0.0f, 1.0f );
+				fAmpl = Clamp( static_cast<float>(fabs( ( fCurWaveStart + fCurWaveLen ) - fCurOffset * 2.0f ) / ( fCurWaveLen - fCurWaveStart )), 0.0f, 1.0f );
 				fAmpl = 1.0f - fAmpl * fAmpl;
 				fAmpl = DEF_WAVE_MIN_AMPL + fAmpl * DEF_WAVE_MAX_AMPL;
 			}
